Change the number of characters in db for IPv6

+++ sql/database_settings.sql
@@ -7,30 +7,30 @@ SET time_zone = "+00:00";
 /*!40101 SET NAMES utf8 */;
 
 CREATE TABLE IF NOT EXISTS `captions` (
-	`id` int(7) NOT NULL AUTO_INCREMENT,
+	`id` int(8) NOT NULL AUTO_INCREMENT,
 	`capt` varchar(256) NOT NULL,
 	`repl` varchar(256) NOT NULL,
-	`user_id` int(7) NOT NULL DEFAULT '0',
+	`user_id` int(8) NOT NULL DEFAULT '0',
 	`client_address` varchar(50) NOT NULL DEFAULT '',
 	`updated` datetime DEFAULT NULL,
 	PRIMARY KEY (id)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
 
 CREATE TABLE IF NOT EXISTS `hints` (
-	`id` int(7) NOT NULL AUTO_INCREMENT,
+	`id` int(8) NOT NULL AUTO_INCREMENT,
 	`tag` varchar(200) NOT NULL,
 	`hint` varchar(4096) NOT NULL,
-	`user_id` int(7) NOT NULL DEFAULT '0',
+	`user_id` int(8) NOT NULL DEFAULT '0',
 	`client_address` varchar(50) NOT NULL DEFAULT '',
 	`updated` datetime DEFAULT NULL,
 	PRIMARY KEY (id)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
 
 CREATE TABLE IF NOT EXISTS `settings` (
-	`id` bigint(8) NOT NULL AUTO_INCREMENT,
+	`id` int(8) NOT NULL AUTO_INCREMENT,
 	`name` tinytext,
 	`value` varchar(512) DEFAULT NULL, 
-	`user_id` int(7) NOT NULL DEFAULT '0',
+	`user_id` int(8) NOT NULL DEFAULT '0',
 	`client_address` varchar(50) NOT NULL DEFAULT '',
 	`updated` datetime DEFAULT NULL,
 	PRIMARY KEY (id)
